requirements.txt
asgiref==3.5.1
Django==3.0.3
djangorestframework==3.13.1
psycopg2-binary==2.8.6
pytz==2022.1
sqlparse==0.4.2
